# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkout:
    post:
      operationId: createCheckout
      tags: [checkout]
      summary: Start a checkout
      description: >
        Starts or resumes a Stripe checkout for one published artwork. Signed-in
        customers may omit email; guests must provide an email address for their
        receipt and for later purchase claiming after they create an account.
      security:
        - clerkSession: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout session created or resumed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Invalid checkout input or a guest email is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Artwork is unavailable for purchase.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The idempotency key or artwork reservation conflicts with an existing checkout.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Stripe could not open checkout.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: The artwork is not fully configured for checkout.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /functions/v1/credit-pack-options:
    get:
      operationId: getCreditPackOptions
      tags: [credit-packs]
      summary: Get standalone credit purchase options
      description: Returns the server-configured unit price and maximum quantity after verifying the Stripe credit price is available.
      security:
        - clerkSession: []
      responses:
        "200":
          description: Purchase options based on the active one-time Stripe price.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditPackOptionsResponse"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: The generation credit price is not configured in Stripe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /functions/v1/credit-pack-checkouts:
    post:
      operationId: createCreditPackCheckout
      tags: [credit-packs]
      summary: Start a standalone credit pack checkout
      description: Creates a Stripe Checkout session for a server-priced quantity of generation credits. It does not grant artwork access.
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditPackCheckoutInput"
      responses:
        "200":
          description: Checkout session created or resumed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditPackCheckoutResponse"
        "400":
          description: Invalid credit quantity or idempotency key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The idempotency key conflicts with an existing checkout.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Stripe could not open checkout.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: The generation credit price is not configured in Stripe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /functions/v1/credit-pack-checkouts/{sessionId}:
    get:
      operationId: getCreditPackCheckoutStatus
      tags: [credit-packs]
      summary: Check a standalone credit checkout after returning from Stripe
      security:
        - clerkSession: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
      responses:
        "200":
          description: Checkout status scoped to the signed-in customer.
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditPackCheckoutStatus"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Credit checkout was not found for this customer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /functions/v1/funnel-events:
    post:
      operationId: recordFunnelEvent
      tags: [sales]
      summary: Record a privacy-safe storefront funnel event
      description: >
        Records an allowlisted product-interest event without customer,
        session, payment, or personal fields. Duplicate event IDs are ignored.
      security:
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FunnelEventRequest"
      responses:
        "200":
          description: Event accepted or already recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunnelEventResponse"
        "400":
          description: Invalid event or unpublished artwork.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Event recording is temporarily unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
...
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [code, message]
...
    CheckoutRequest:
      type: object
      properties:
        artworkId:
          type: string
          minLength: 1
          maxLength: 200
        idempotencyKey:
          type: string
          format: uuid
        selectedPreviewId:
          type: [string, "null"]
          minLength: 1
          maxLength: 200
        email:
          type: [string, "null"]
          format: email
          maxLength: 320
      required: [artworkId, idempotencyKey]
    CheckoutResponse:
      type: object
      properties:
        purchaseId:
          type: string
        checkoutUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
        includedCredits:
          type: integer
          minimum: 0
      required: [purchaseId, checkoutUrl, expiresAt, includedCredits]
    CreditPackOptionsResponse:
      type: object
      additionalProperties: false
      properties:
        currency:
          type: string
        creditPriceCents:
          type: integer
          minimum: 1
        maxCredits:
          type: integer
          minimum: 1
      required: [currency, creditPriceCents, maxCredits]
    CreditPackCheckoutInput:
      type: object
      additionalProperties: false
      properties:
        credits:
          type: integer
          minimum: 1
          maximum: 50
        idempotencyKey:
          type: string
          format: uuid
      required: [credits, idempotencyKey]
    CreditPackCheckoutResponse:
      type: object
      additionalProperties: false
      properties:
        creditPackPurchaseId:
          type: string
        checkoutUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
        credits:
          type: integer
          minimum: 1
        amountCents:
          type: integer
          minimum: 1
      required:
        - creditPackPurchaseId
        - checkoutUrl
        - expiresAt
        - credits
        - amountCents
    CreditPackCheckoutStatus:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [reserved, paid, expired, refunded]
        credits:
          type: integer
          minimum: 1
      required: [status, credits]
...
    FunnelEventRequest:
      type: object
      properties:
        eventId:
          type: string
          format: uuid
        eventType:
          type: string
          enum: [product_viewed]
        artworkId:
          type: string
          minLength: 1
          maxLength: 200
      required: [eventId, eventType, artworkId]
    FunnelEventResponse:
      type: object
      properties:
        recorded:
          type: boolean
      required: [recorded]
...