# lib/api-zod/src/generated/types/creditPackCheckoutInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreditPackCheckoutInput {
  /**
     * @minimum 1
     * @maximum 50
     */
  credits: number;
  idempotencyKey: string;
}