# lib/api-zod/src/generated/types/creditPackCheckoutResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreditPackCheckoutResponse {
  creditPackPurchaseId: string;
  checkoutUrl: string;
  expiresAt: Date;
  /** @minimum 1 */
  credits: number;
  /** @minimum 1 */
  amountCents: number;
}