# lib/api-zod/src/generated/types/creditPackCheckoutStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreditPackCheckoutStatusStatus } from './creditPackCheckoutStatusStatus';

export interface CreditPackCheckoutStatus {
  status: CreditPackCheckoutStatusStatus;
  /** @minimum 1 */
  credits: number;
}