# lib/api-zod/src/generated/types/creditPackOptionsResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreditPackOptionsResponse {
  currency: string;
  /** @minimum 1 */
  creditPriceCents: number;
  /** @minimum 1 */
  maxCredits: number;
}